+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Added
+
+- New command line that allows to hide the window when running PulseEffects in service mode. This will allow tray
+  applets to hide our window.
+
 ## [4.7.3]
 
 ### Added